+++ app/Http/Controllers/MascotaClienteController.php
@@ -15,23 +15,36 @@ class MascotaClienteController extends Controller
 
         $mascota->clientes()->attach($request->rut_cliente);
 
-        return redirect()->back()->with("success","");
+        return redirect()->back()->with("success", "");
     }
 
     public function destroy(Request $request)
     {
         $mascota = Mascota::where("id", $request->id_mascota)->first();
 
+        $cantidadDuennos = $mascota->clientes()->count();
+
+        if ($cantidadDuennos <= 1) {
+            return redirect()->back()->with("error", "La mascota no puede quedar sin dueños, por lo que no se puede desligar.");
+        }
+
         $mascota->clientes()->detach($request->rut_cliente);
 
-        return redirect()->back()->with("success","");
+        return redirect()->back()->with("success", "");
     }
 
     public function destroy2(Request $request, Mascota $mascota)
-{
-    $mascota->clientes()->detach($request->rut_cliente);
-    // Redireccionar con un mensaje de éxito
-    return redirect()->back()->with('success', 'La mascota ha sido desligada del cliente con éxito.');
-}
+    {
+
+        $cantidadDuennos = $mascota->clientes()->count();
+
+        if ($cantidadDuennos <= 1) {
+            return redirect()->back()->with("error", "La mascota no puede quedar sin dueños, por lo que no se puede desligar.");
+        }
+
+        $mascota->clientes()->detach($request->rut_cliente);
+        // Redireccionar con un mensaje de éxito
+        return redirect()->back()->with('success', 'La mascota ha sido desligada del cliente con éxito.');
+    }
 
 }
